Add files via upload
working flask server with sender and receiver scripts and test images. The server must sit in /home/ubuntu/seefood-master/, i.e. in the parent directory of the saved_model folder. Also a 'temp' folder is required: /home/ubuntu/seefood-master/temp/, exactly like that. To get it working on your own ec2, you need to open port 5000 under the Security Groups > select your ec2 > bottom of page select "Inbound" and then "create new rule" and set it to how I have it in the image. Also in the test scripts, you need to change the ip4, obviously and make sure you have all the libraries. If you are running the ec2 from the template given to us, no installs are necessary. Make sure to make a snapshot of your ec2 just in case!!!